FU-54 refactor: 인증 관련 클래스를 auth 패키지 하위로 이동
